max activations working

+++ indl/cnn_helper_1D.py
@@ -8,6 +8,7 @@ Results example: ![Visualization](http://i.imgur.com/4nj4KjN.jpg)
 
 import time
 import numpy as np
+import tensorflow as tf
 from tensorflow.keras import layers
 from tensorflow.keras import backend as K
 
@@ -70,13 +71,13 @@ def process_timeseries(x, former):
 def visualize_layer(model,
                     layer_idx,
                     step=1.,
-                    epochs=15,
+                    epochs=200,
                     upscaling_steps=9,
                     upscaling_factor=1.2,
-                    output_dim=(412, 412),
+                    output_dim=(801, 58),
                     filter_range=(0, None)):
     """Visualizes the most relevant filters of one conv-layer in a certain model.
-
+    https://github.com/keras-team/keras/blob/master/examples/conv_filter_visualization.py
     # Arguments
         model: The model containing layer_name.
         layer_name: The name of the layer to be visualized.
@@ -94,7 +95,7 @@ def visualize_layer(model,
                       the last filter will be inferred as the upper boundary.
     """
 
-    def _generate_filter_dat(input_dat,
+    def _generate_filter_dat(output_shape,
                              layer_output,
                              filter_index):
         """Generates image for one particular filter.
@@ -116,51 +117,57 @@ def visualize_layer(model,
         @tf.function
         def iterate(train_dat):
             with tf.GradientTape() as tape:
+                tape.watch(train_dat)
                 predictions = model(train_dat)
                 loss = loss_object(predictions)
-            gradients = tape.gradient(loss, model.trainable_variables)
+            gradients = tape.gradient(loss, train_dat)
             # normalization trick: we normalize the gradient
             gradients = normalize(gradients)
             return loss, gradients
 
         # we start with some random noise
-        intermediate_dim = tuple(
-            int(x / (upscaling_factor ** upscaling_steps)) for x in output_dim)
-
-        input_dat = np.random.random((1, intermediate_dim[0], intermediate_dim[1])).astype(np.float32)
+        # intermediate_dim = tuple(
+        #     int(x / (upscaling_factor ** upscaling_steps)) for x in output_dim)
+        #
+        # input_dat = np.random.random((1, intermediate_dim[0], intermediate_dim[1])).astype(np.float32)
 
         # Slowly upscaling towards the original size prevents
         # a dominating high-frequency of the to visualized structure
         # as it would occur if we directly compute the 412d-image.
         # Behaves as a better starting point for each following dimension
         # and therefore avoids poor local minima
+
+        input_data = tf.convert_to_tensor(np.random.randn(*output_shape).astype(np.float32)[None, :, :])
+
         for up in reversed(range(upscaling_steps)):
             # we run gradient ascent for e.g. 20 steps
             for _ in range(epochs):
-                loss_value, grads_value = iterate([input_dat])
-                input_dat += grads_value * step
+                loss_value, grads_value = iterate([input_data])
+                input_data += grads_value * step
 
                 # some filters get stuck to 0, we can skip them
                 if loss_value <= K.epsilon():
                     return None
 
-            # Calulate upscaled dimension
-            intermediate_dim = tuple(
-                int(x / (upscaling_factor ** up)) for x in output_dim)
-            # TODO: Upscale
+            if False:
+                # Calulate upscaled dimension
+                intermediate_dim = tuple(
+                    int(x / (upscaling_factor ** up)) for x in output_dim)
+                # TODO: Upscale
+                dat = deprocess_timeseries(input_dat[0])
+                # TODO: Reshape
+                input_dat = [process_timeseries(dat, input_dat[0])]
+
+        if False:
+            # decode the resulting input image
             dat = deprocess_timeseries(input_dat[0])
-            # TODO: Reshape
-            input_dat = [process_timeseries(dat, input_dat[0])]
+            e_time = time.time()
+            print('Costs of filter {:3}: {:5.0f} ( {:4.2f}s )'.format(filter_index,
+                                                                      loss_value,
+                                                                      e_time - s_time))
+        return input_data[0], loss_value
 
-        # decode the resulting input image
-        dat = deprocess_timeseries(input_dat[0])
-        e_time = time.time()
-        print('Costs of filter {:3}: {:5.0f} ( {:4.2f}s )'.format(filter_index,
-                                                                  loss_value,
-                                                                  e_time - s_time))
-        return dat, loss_value
-
-    def _draw_filters(filters, n=None):
+    def _stitch_filters(max_acts, n=None):
         """Draw the best filters in a nxn grid.
 
         # Arguments
@@ -170,15 +177,13 @@ def visualize_layer(model,
                If none, the largest possible square will be used
         """
         if n is None:
-            n = int(np.floor(np.sqrt(len(filters))))
+            n = int(np.floor(np.sqrt(len(max_acts))))
 
         # the filters that have the highest loss are assumed to be better-looking.
         # we will only keep the top n*n filters.
-        filters.sort(key=lambda x: x[1], reverse=True)
-        filters = filters[:n * n]
+        max_acts.sort(key=lambda x: x[1], reverse=True)
+        max_acts = max_acts[:n * n]
 
-        # build a black picture with enough space for
-        # e.g. our 8 x 8 filters of size 412 x 412, with a 5px margin in between
         MARGIN = 5
         width = n * output_dim[0] + (n - 1) * MARGIN
         height = n * output_dim[1] + (n - 1) * MARGIN
@@ -187,46 +192,36 @@ def visualize_layer(model,
         # fill the picture with our saved filters
         for i in range(n):
             for j in range(n):
-                img, _ = filters[i * n + j]
+                img, _ = max_acts[i * n + j]
                 width_margin = (output_dim[0] + MARGIN) * i
                 height_margin = (output_dim[1] + MARGIN) * j
                 stitched_filters[
                     width_margin: width_margin + output_dim[0],
                     height_margin: height_margin + output_dim[1], :] = img
 
-    # this is the placeholder for the input data
-    assert len(model.inputs) == 1
-    input_dat = model.inputs[0]
-
     output_layer = model.layers[layer_idx]
-    assert isinstance(output_layer, tf.keras.layers.Conv1D)
-
-    # Compute to be processed filter range
-    filter_lower = filter_range[0]
-    filter_upper = (filter_range[1]
-                    if filter_range[1] is not None
-                    else len(output_layer.get_weights()[1]))
-    assert(0 <= filter_lower < filter_upper <= len(output_layer.get_weights()[1]))
-    print('Compute filters {:} to {:}'.format(filter_lower, filter_upper))
-
+    max_filts = len(output_layer.get_weights()[1])
+    max_filts = max_filts if filter_range[1] is None else min(max_filts, filter_range[1])
     # iterate through each filter and generate its corresponding time series
-    processed_filters = []
-    for f in range(filter_lower, filter_upper):
-        dat_loss = _generate_filter_dat(input_dat, output_layer.output, f)
+    maximizing_activations = []
+    for f_ix in range(filter_range[0], max_filts):
+        if isinstance(output_layer, tf.keras.layers.Conv1D):
+            model_output = output_layer.output[:, :, f_ix]
+        else:
+            model_output = output_layer.output[f_ix]
+        max_model = tf.keras.Model(model.input, model_output)
+        max_data, loss_vals = _generate_filter_dat(output_dim, max_model)
+        if max_data is not None:
+            maximizing_activations.append(max_data)
 
-        if dat_loss is not None:
-            processed_filters.append(dat_loss)
-
-    print('{} filter processed.'.format(len(processed_filters)))
-    # Finally draw and store the best filters to disk
-    _draw_filters(processed_filters)
+    print('{} filter processed.'.format(len(maximizing_activations)))
+    # Stitch timeseries together into one mega timeseries with NaN gaps.
+    _stitch_filters(maximizing_activations)
 
 
 if __name__ == '__main__':
     from pathlib import Path
     import os
-    import tempfile
-    import tensorflow as tf
     from tensorflow.keras.models import load_model
 
     if Path.cwd().stem == 'indl':
@@ -235,17 +230,19 @@ if __name__ == '__main__':
     model_file = Path.cwd() / 'data' / 'kjm_ecog' / 'converted' / 'faces_basic' / 'mv_model_full.h5'
     model = load_model(str(model_file))
 
-    model.layers[-1].activation = tf.keras.activations.linear
-    # Save and load the model to actually apply the change.
-    tmp_path = Path(tempfile.gettempdir()) / (next(tempfile._get_candidate_names()) + '.h5')
-    try:
-        model.save(str(tmp_path))
-        model = load_model(str(tmp_path))
-    finally:
-        tmp_path.unlink()
+    if False:
+        import tempfile
+        model.layers[-1].activation = tf.keras.activations.linear
+        # Save and load the model to actually apply the change.
+        tmp_path = Path(tempfile.gettempdir()) / (next(tempfile._get_candidate_names()) + '.h5')
+        try:
+            model.save(str(tmp_path))
+            model = load_model(str(tmp_path))
+        finally:
+            tmp_path.unlink()
     model.summary()
 
-    layer_idx = -7
-    visualize_layer(model, layer_idx, upscaling_steps=1,
-                    upscaling_factor=1.0,
+    layer_idx = 14  # [2, 6, 10, 14]
+    visualize_layer(model, layer_idx,
+                    upscaling_steps=1, upscaling_factor=1.0,
                     output_dim=model.get_input_shape_at(0)[1:])
